pyproject.toml
[project]
name = "paper-digest"
version = "0.1.0"
description = "Daily arXiv paper digest with local LLM scoring and Obsidian vault chat"
requires-python = ">=3.12"
dependencies = [
    "marker-pdf>=1.10.2",
    "ollama>=0.6.1",
    "requests>=2.33.1",
]

[project.scripts]
run-digest = "digest.run:main"
convert-pdf = "digest.convert:main"
vault-chat = "vault_chat.chat:main"
